no click builtin, rm gdrive, consolidate getGooseInstallLink and MCPServer type
